diskview
Added the diskview package. Diskview is the layer which talks to the
file system and returns the mmap region to the user. It also has a
builtin LRU cache for fast reads on frequent pages. The locking
mechanism in the package is not refined and will be worked on next.